CODE CLEANING: ML & PLOTTING
 - cleaning some of the ML functions (+ bug fixing)
 - cleaning some plotting main functions.
 - bug fixing
