+++ verticapy/learn/metrics/__init__.py
@@ -34,7 +34,6 @@ from verticapy.machine_learning.metrics.classification import (
     specificity_score,
 )
 from verticapy.machine_learning.metrics.regression import (
-    aic_bic,
     aic_score,
     bic_score,
     anova_table,
